requirements.txt
python~=3.9.7
tensorflow~=2.9.0
keras~=2.9.0
numpy~=1.22.4
Flask~=2.1.2
opencv-python~=4.5.5.64
eventlet~=0.33.1
pandas~=1.4.2
scikit-learn~=1.1.1
matplotlib~=3.5.2
Pillow~=9.1.1
python-engineio~=3.13.2
python-socketio~=4.6.1